Add files via upload
Day 1 of 160  

Hello #connections #gfg160 #geekstreak2024
I'm excited to announce that I've joined the GFG160 challenge! For the next 160 days, I'll be diving deep into data structures and algorithms, solving daily coding problems, and pushing my problem-solving skills to new heights.
Problem - Second Largest
Given an array of positive integers arr[], return the second largest element from the array. If the second largest element doesn't exist then return -1.
Note: The second largest element should not be equal to the largest element.
Examples:
Input: arr[] = [12, 35, 1, 10, 34, 1]
Output: 34
Explanation: The largest element of the array is 35 and the second largest element is 34.
Input: arr[] = [10, 5, 10]
Output: 5
Explanation: The largest element of the array is 10 and the second largest element is 5.
Input: arr[] = [10, 10, 10]
Output: -1
Explanation: The largest element of the array is 10 and the second largest element does not exist.
#gfg160 #geekstreak2024 #geeksforgeeks